Allow custom transfer of table
The default method blindly transfer everything. This is not always a good way.
Think of an object person and its class. Transferring everything means
that the class will be transfered every time a person is requested.
Since the class doesn't change over time, this create many redundancies.

If the application using the transfer method can handle a proper analysis of
what to transfer, the situation above won't appear. Allowing a custom
copy of tables makes it possible to analyze the data and make a minimal
transfer in the same loop.