Fix Table looping issue
one possibilities in Lua is to have a table referencing itself at some
point (e.g. a car table having a wheel table referencing its parent
table). This would have resulted in an infinite cycle in the previous
implementation.

Now, for tables, the system take into account all the copied table up to
that point. This also means that table that are used in multiple places
(e.g. the car table having two identical wheel tables) will not be
duplicated.